# BE/Laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/search/": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Tìm kiếm",
                "description": "Tìm kiếm người dùng hoặc bài post",
                "operationId": "search",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/search-user/{keyword}": {
            "post": {
                "tags": [
                    "Search"
                ],
                "summary": "Tìm kiếm người dùng",
                "description": "Tìm kiếm người dùng trong phần tin nhắn",
                "operationId": "08920553db3cd583360da2ffc28b5e46",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "header",
                        "description": "Điền email hoặc tên",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "405": {
                        "description": "Method Not Allowed"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}